+++ lib/instruction_execute/instruction_execute.ml
@@ -3,8 +3,8 @@ open Hardcaml.Signal
 
 module I = struct
   type 'a t = {
-    rs_val : 'a; [@bits 32]
-    rt_val : 'a; [@bits 32]
+    alu_a : 'a; [@bits 32]
+    alu_b : 'a; [@bits 32]
     imm : 'a; [@bits 32]
     alu_control : 'a; [@bits (width Control_unit.Alu_ops.default)]
     sel_shift_for_alu: 'a;
@@ -19,8 +19,8 @@ end
 
 let circuit_impl (scope : Scope.t) (input : _ I.t) =
   let shift_amount = uresize input.imm.:[10, 6] 32 in
-  let arg_a = mux2 input.sel_shift_for_alu shift_amount input.rs_val in
-  let arg_b = mux2 input.sel_imm_for_alu input.imm input.rt_val in
+  let arg_a = mux2 input.sel_shift_for_alu shift_amount input.alu_a in
+  let arg_b = mux2 input.sel_imm_for_alu input.imm input.alu_b in
   let alu = Alu.hierarchical scope { Alu.I.arg_a; arg_b; alu_control = input.alu_control} in
   { O.alu_result = alu.result }